Refactored truly common from default-sc test functions
