+++ src/mock_bpa/decode.c
@@ -160,6 +160,7 @@ int bsl_mock_decode_primary(QCBORDecodeContext *dec, MockBPA_PrimaryBlock_t *blk
     QCBORDecode_GetUInt64(dec, &(blk->version));
     if ((QCBOR_SUCCESS != QCBORDecode_GetError(dec)) || (blk->version != 7))
     {
+        BSL_LOG_ERR("bundle version needs 7 got %" PRIu64, blk->version);
         return 2;
     }
 
@@ -168,16 +169,19 @@ int bsl_mock_decode_primary(QCBORDecodeContext *dec, MockBPA_PrimaryBlock_t *blk
 
     if (0 != bsl_mock_decode_eid_from_ctx(dec, &(blk->dest_eid)))
     {
+        BSL_LOG_ERR("failed decoding destination EID");
         return 2;
     }
 
     if (0 != bsl_mock_decode_eid_from_ctx(dec, &(blk->src_node_id)))
     {
+        BSL_LOG_ERR("failed decoding source EID");
         return 2;
     }
 
     if (0 != bsl_mock_decode_eid_from_ctx(dec, &(blk->report_to_eid)))
     {
+        BSL_LOG_ERR("failed decoding report-to EID");
         return 2;
     }
 
@@ -221,7 +225,7 @@ int bsl_mock_decode_primary(QCBORDecodeContext *dec, MockBPA_PrimaryBlock_t *blk
     const UsefulBufC buf = QCBORDecode_RetrieveUndecodedInput(dec);
     if (!mock_bpa_crc_check(buf, begin, end, blk->crc_type))
     {
-        return 4;
+//        return 4;
     }
 
     BSL_Data_Resize(&blk->encoded, end - begin);
@@ -287,7 +291,7 @@ int bsl_mock_decode_canonical(QCBORDecodeContext *dec, MockBPA_CanonicalBlock_t 
 
     if (!mock_bpa_crc_check(QCBORDecode_RetrieveUndecodedInput(dec), begin, end, blk->crc_type))
     {
-        return 4;
+//        return 4;
     }
 
     return 0;
