+++ test/test_DynamicMemCbs.c
@@ -30,7 +30,7 @@
 #include <mock_bpa/agent.h>
 #include <mock_bpa/log.h>
 
-#include "bsl_test_utils.h"
+#include "DefaultScUtils.h"
 
 static int malloc_cnt  = 0;
 static int realloc_cnt = 0;
@@ -96,7 +96,8 @@ void _setUp(void)
 {
     setenv("BSL_TEST_LOCAL_IPN_EID", "ipn:2.1", 1);
     BSL_CryptoInit();
-    TEST_ASSERT_EQUAL(0, BSL_TestContext_Init(&LocalTestCtx, true));
+    TEST_ASSERT_EQUAL(0, BSL_TestContext_Init(&LocalTestCtx));
+    BSL_TestUtils_SetupDefaultSecurityContext(&LocalTestCtx.bsl);
 
     BSL_SecParam_Init(&param_aes_variant_128);
     BSL_SecParam_Init(&param_use_wrap_key);
